requirements.txt
# File parsing
PyPDF2>=3.0.0
python-docx>=1.1.0

# LLM integration (Gemini)
google-generativeai>=0.8.0

# Environment variable management
python-dotenv>=1.0.0

# PDF generation (for creating sample resumes)
reportlab>=4.0.0

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0
